refactor: group embed-viz and HITL MCP-App code into feature folders
Split the flat src/web/apps/src/lib/ directory into feature-scoped folders so
each browser-side MCP-App bundle is self-contained:

- embed/  — Tableau viz embedding for get-view/get-workbook (mcp-app entry,
            handleToolResult, embedTableauViz, loadTableauEmbeddingApi,
            getEmbedTokenToolClient, openInTableauLink) + mcp-app.html
- hitl/   — in-iframe confirm panel for delete/update preview tools
            (hitl-confirm entry, handleConfirmResult, *ConfirmClient renderers)
            + hitl-confirm.html
- shared/ — mcp-app.css, showError, vite-env.d.ts, assets/, and a new
            vizContainer.ts holding TABLEAU_VIZ_CONTAINER_ID (extracted to
            keep shared/ free of a shared->embed dependency)

HTML templates move next to their entry .ts with relative <script src>. Each
Vite build's root is set to its feature folder in build.ts so single-file
output still lands flat as dist/mcp-app.html and dist/hitl-confirm.html;
appConfig.ts and server.web.ts are unchanged. All moves use git mv to preserve
history. 2365 unit tests pass, lint clean, build produces flat dist.
