+++ src/web/apps/src/embed/getEmbedTokenToolClient.test.ts
