+++ src/web/apps/src/embed/getEmbedTokenToolClient.ts
