+++ src/web/apps/src/embed/handleToolResult.test.ts
