+++ src/web/apps/src/embed/mcp-app.html
@@ -9,6 +9,6 @@
   <main class="main">
     <div id="tableauVizContainer" class="viz-container"></div>
   </main>
-  <script type="module" src="/src/hitl-confirm.ts"></script>
+  <script type="module" src="./mcp-app.ts"></script>
 </body>
 </html>
