+++ src/web/apps/src/embed/mcp-app.ts
@@ -1,11 +1,11 @@
-import './mcp-app.css';
+import '../shared/mcp-app.css';
 
 import { App } from '@modelcontextprotocol/ext-apps';
 import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 
 import pkg from '~/package.json';
 
-import { handleToolResult } from './lib/handleToolResult.js';
+import { handleToolResult } from './handleToolResult.js';
 
 const app = new App({ name: 'Tableau MCP App', version: pkg.version });
 app.ontoolresult = (result: CallToolResult) => {
