+++ src/web/apps/src/hitl/deleteWorkbookConfirmClient.ts
