+++ src/web/apps/src/hitl/handleConfirmResult.ts
@@ -1,6 +1,7 @@
 import type { App } from '@modelcontextprotocol/ext-apps';
 import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 
+import { showError } from '../shared/showError.js';
 import {
   isDeleteDatasourceConfirmResult,
   renderDeleteDatasourceConfirm,
@@ -13,7 +14,6 @@ import {
   isDeleteWorkbookConfirmResult,
   renderDeleteWorkbookConfirm,
 } from './deleteWorkbookConfirmClient.js';
-import { showError } from './showError.js';
 import {
   isUpdateCloudExtractRefreshTaskConfirmResult,
   renderUpdateCloudExtractRefreshTaskConfirm,
