+++ src/web/apps/src/hitl/hitl-confirm.html
@@ -9,6 +9,6 @@
   <main class="main">
     <div id="tableauVizContainer" class="viz-container"></div>
   </main>
-  <script type="module" src="/src/mcp-app.ts"></script>
+  <script type="module" src="./hitl-confirm.ts"></script>
 </body>
 </html>
