+++ src/web/apps/src/shared/mcp-app.css
