+++ src/web/apps/src/shared/showError.ts
@@ -1,5 +1,5 @@
-import DISCONNECTED_SVG from '../assets/disconnected.svg?raw';
-import { TABLEAU_VIZ_CONTAINER_ID } from './embedTableauViz.js';
+import DISCONNECTED_SVG from './assets/disconnected.svg?raw';
+import { TABLEAU_VIZ_CONTAINER_ID } from './vizContainer.js';
 
 export type Scenario = 'TOOL_ERROR' | 'PARSE_ERROR' | 'AUTH_ERROR' | 'EMBED_LOAD_ERROR';
