+++ src/web/apps/src/shared/vite-env.d.ts
