+++ src/web/apps/src/shared/vizContainer.ts
@@ -1,0 +1,6 @@
+/**
+ * @file Shared identifier for the app iframe's viz container element.
+ * The id of the <div> present in both app HTML templates; imported by the embed logic
+ * (to mount the viz) and by the shared error UI (to render into the same container).
+ */
+export const TABLEAU_VIZ_CONTAINER_ID = 'tableauVizContainer';
